+++ Source/Rendering/Text/FTInputLabel.cpp
@@ -146,7 +146,7 @@ void FTInputLabel::updateCursorRendererPos() const {
 
 void FTInputLabel::onPressed(FTButton* button, const FTMouseButtonPressedEvent& event) {
     auto pos = convertMouseToLocalCoordinates(event.cursor_pos_);
-    for (int i = 0; i < input_text_.size(); i++) {
+    for (size_t i = 0; i < input_text_.size(); i++) {
         auto glyph = label_->getMeshData()->getGlyphBounds(i);
         auto delta1 = pos.x - glyph.first;
         auto delta2 = glyph.second - pos.x;
